+++ source/win-assets/CustomCSS.vue
@@ -99,6 +99,5 @@ function saveCSS (): void {
 }
 </script>
 
-<style lang="less">
-//
+<style lang="css" scoped>
 </style>