+++ source/win-assets/SnippetsTab.vue
@@ -288,6 +288,5 @@ function openSnippetsDirectory (): void {
 }
 </script>
 
-<style lang="less">
-//
+<style lang="css" scoped>
 </style>