+++ source/win-onboarding/pages/WritingCheckPage.vue
@@ -87,5 +87,5 @@ function activateLanguageTool () {
 }
 </script>
 
-<style lang="less">
+<style lang="css" scoped>
 </style>